CalendarScreen.kt: Locale according to app Locale chosen for calendar. There is also "Selected*" fields calculation fixed. EditWorkDayViewModel.kt: Lunch time default set to "00:00". IntroScreen.kt: "Continue" button size increased. Saving project to db fixed: Unique "date" + "project_name" columns in "project" table. SettingsScreen.kt: "Work type" drop down now shows first item by default.

+++ app/src/main/java/com/akiwiksten/worktime30/domain/SaveProjectsUseCase.kt
@@ -3,26 +3,39 @@ package com.akiwiksten.worktime30.domain
 import com.akiwiksten.worktime30.core.ZERO_TIME
 import com.akiwiksten.worktime30.data.database.entity.ProjectEntity
 import com.akiwiksten.worktime30.data.database.entity.ProjectNameEntity
+import com.akiwiksten.worktime30.data.database.entity.WorkDayEntity
 import com.akiwiksten.worktime30.data.repository.ProjectRepository
+import com.akiwiksten.worktime30.data.repository.WorkDayRepository
 import javax.inject.Inject
 
 class SaveProjectsUseCase @Inject constructor(
-    private val projectRepository: ProjectRepository
+    private val projectRepository: ProjectRepository,
+    private val workDayRepository: WorkDayRepository
 ) {
     suspend operator fun invoke(
         date: String,
         projectsToSave: List<ProjectEntity>,
-        projectNamesToDelete: List<String>
+        projectNamesToDelete: List<String>,
+        workDayToSave: WorkDayEntity? = null
     ) {
         projectsToSave.forEach { project ->
             projectRepository.insertProjectName(ProjectNameEntity(project.projectName))
             projectRepository.insertProject(project)
         }
+        
+        workDayToSave?.let {
+            workDayRepository.insertWorkDay(it)
+        }
+        
         projectNamesToDelete.forEach { name ->
             projectRepository.deleteProject(
                 ProjectEntity(date = date, projectName = name, projectTime = ZERO_TIME)
             )
-            projectRepository.deleteProjectName(ProjectNameEntity(name))
+            workDayRepository.deleteWorkDay(WorkDayEntity(date = date, projectName = name))
+            
+            if (!projectRepository.isProjectNameUsed(name)) {
+                projectRepository.deleteProjectName(ProjectNameEntity(name))
+            }
         }
     }
 }
